+++ src/test/java/kitchenpos/ordertable/presentation/TableRestControllerTest.java
@@ -4,9 +4,9 @@ import io.restassured.RestAssured;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.common.controller.ControllerTest;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
-import kitchenpos.domain.repository.TableGroupRepository;
+import kitchenpos.tablegroup.domain.repository.TableGroupRepository;
 import kitchenpos.ordertable.presentation.dto.EmptyChangeRequest;
 import kitchenpos.ordertable.presentation.dto.GuestChangeRequest;
 import org.junit.jupiter.api.Test;
